[feat] 프로토콜에 대한 메서드를 extension으로 구현
