+++ Quick-Kick/Managers/UserDefaultsManager.swift
@@ -5,7 +5,6 @@
 //  Created by 반성준 on 12/18/24.
 //
 
-// UserDefaultsManager.swift
 import Foundation
 
 // MARK: - Protocol
@@ -15,6 +14,43 @@ protocol UserDataManageable {
     var isLoggedIn: Bool { get set }
     var autoLoginOption: Bool { get set }
     var rememberIDOption: Bool { get set }
+    func setLoggedOut()
+}
+
+// MARK: - Protocol Extension
+extension UserDataManageable {
+    func saveUser(_ user: User) {
+        if let encoded = try? JSONEncoder().encode(user) {
+            UserDefaults.standard.set(encoded, forKey: "user")
+        }
+    }
+
+    func getUser() -> User? {
+        if let data = UserDefaults.standard.data(forKey: "user"),
+           let user = try? JSONDecoder().decode(User.self, from: data) {
+            return user
+        }
+        return nil
+    }
+
+    func setLoggedOut() {
+        UserDefaults.standard.set(false, forKey: "LoginStatus")
+    }
+
+    var isLoggedIn: Bool {
+        get { UserDefaults.standard.bool(forKey: "LoginStatus") }
+        set { UserDefaults.standard.set(newValue, forKey: "LoginStatus") }
+    }
+
+    var autoLoginOption: Bool {
+        get { UserDefaults.standard.bool(forKey: "AutoLoginOption") }
+        set { UserDefaults.standard.set(newValue, forKey: "AutoLoginOption") }
+    }
+
+    var rememberIDOption: Bool {
+        get { UserDefaults.standard.bool(forKey: "RememberIDOption") }
+        set { UserDefaults.standard.set(newValue, forKey: "RememberIDOption") }
+    }
 }
 
 // MARK: - User Struct
@@ -28,41 +64,4 @@ struct User: Codable {
 final class UserDefaultsManager: UserDataManageable {
     static let shared = UserDefaultsManager()
     private init() {}
-
-    private enum Keys {
-        static let user = "user"
-        static let loginStatus = "LoginStatus"
-        static let autoLogin = "AutoLoginOption"
-        static let rememberID = "RememberIDOption"
-    }
-
-    // MARK: - UserDataManageable
-    func saveUser(_ user: User) {
-        if let encoded = try? JSONEncoder().encode(user) {
-            UserDefaults.standard.set(encoded, forKey: Keys.user)
-        }
-    }
-
-    func getUser() -> User? {
-        if let data = UserDefaults.standard.data(forKey: Keys.user),
-           let user = try? JSONDecoder().decode(User.self, from: data) {
-            return user
-        }
-        return nil
-    }
-
-    var isLoggedIn: Bool {
-        get { UserDefaults.standard.bool(forKey: Keys.loginStatus) }
-        set { UserDefaults.standard.set(newValue, forKey: Keys.loginStatus) }
-    }
-
-    var autoLoginOption: Bool {
-        get { UserDefaults.standard.bool(forKey: Keys.autoLogin) }
-        set { UserDefaults.standard.set(newValue, forKey: Keys.autoLogin) }
-    }
-
-    var rememberIDOption: Bool {
-        get { UserDefaults.standard.bool(forKey: Keys.rememberID) }
-        set { UserDefaults.standard.set(newValue, forKey: Keys.rememberID) }
-    }
 }
